Converting all tests to new structure
