+++ addon/index.js
@@ -1,4 +1,7 @@
-export { runTask, scheduleTask, throttleTask, cancelTask } from './utils/tasks';
+export { runTask, scheduleTask, throttleTask, cancelTask } from './run-task';
 export { pollTask, pollTaskFor, setShouldPoll, cancelPoll } from './poll-task';
 export { debounceTask, cancelDebounce } from './debounce-task';
 export { runDisposables } from './utils/disposable';
+export { ContextBoundTasksMixin } from './mixins/run';
+export { ContextBoundEventListenersMixin } from './mixins/dom';
+export { DisposableMixin } from './mixins/disposable';
